refactor: introduce `AbstractWALChannel`
